+++ AudioKit/AudioKit/Operations/Signal Modifiers/Filters/Butterworth Filters/AKBandRejectButterworthFilter/AKBandRejectButterworthFilterDSPKernel.hpp
@@ -46,11 +46,11 @@ public:
     void setParameter(AUParameterAddress address, AUValue value) {
         switch (address) {
             case centerFrequencyAddress:
-                centerFrequencyRamper.set(clamp(value, 12.0f, 20000.0f));
+                centerFrequencyRamper.set(clamp(value, (float)12.0, (float)20000.0));
                 break;
 
             case bandwidthAddress:
-                bandwidthRamper.set(clamp(value, 0.0f, 20000.0f));
+                bandwidthRamper.set(clamp(value, (float)0.0, (float)20000.0));
                 break;
 
         }
@@ -71,11 +71,11 @@ public:
     void startRamp(AUParameterAddress address, AUValue value, AUAudioFrameCount duration) override {
         switch (address) {
             case centerFrequencyAddress:
-                centerFrequencyRamper.startRamp(clamp(value, 12.0f, 20000.0f), duration);
+                centerFrequencyRamper.startRamp(clamp(value, (float)12.0, (float)20000.0), duration);
                 break;
 
             case bandwidthAddress:
-                bandwidthRamper.startRamp(clamp(value, 0.0f, 20000.0f), duration);
+                bandwidthRamper.startRamp(clamp(value, (float)0.0, (float)20000.0), duration);
                 break;
 
         }
